+++ projects/ids-enterprise-ng/src/lib/datepicker/soho-datepicker.component.ts
@@ -583,7 +583,7 @@ export class SohoDatePickerComponent extends BaseControlValueAccessor<string | n
     if (this.datepicker) {
       // The processing is required to ensure we use the correct format
       // in the control.
-      if (typeof value === 'string' && this._options.range?.useRange) {
+      if (typeof value === 'string' && value.length > 0 && this._options.range?.useRange) {
         const dates = value.split('-');
         const startValue = Soho.Locale.parseDate(dates[0].trim()) || Soho.Locale.parseDate(dates[0].trim(), { pattern: this._options?.dateFormat });
         const endValue = Soho.Locale.parseDate(dates[1].trim()) || Soho.Locale.parseDate(dates[1].trim(), { pattern: this._options?.dateFormat });